+++ src/__tests__/config.test.ts
@@ -1,4 +1,4 @@
-import { EthrDID } from '..'
+import { EthrDID } from '../index'
 
 describe('configuration', () => {
   it('can use rpcUrl - github #64', () => {