+++ src/__tests__/index.test.ts
@@ -6,6 +6,8 @@ import { createProvider, sleep } from './testUtils'
 import DidRegistryContract from 'ethr-did-registry'
 import { verifyJWT } from 'did-jwt'
 
+import { jest } from '@jest/globals'
+
 jest.setTimeout(30000)
 
 describe('EthrDID', () => {
@@ -23,7 +25,7 @@ describe('EthrDID', () => {
   const provider = createProvider()
 
   beforeAll(async () => {
-    const factory = ContractFactory.fromSolidity(DidRegistryContract).connect(await provider.getSigner(0))
+    const factory = ContractFactory.fromSolidity(EthereumDIDRegistry).connect(await provider.getSigner(0))
 
     let registryContract: Contract
     registryContract = await factory.deploy()
@@ -831,6 +833,24 @@ describe('EthrDID (Meta Transactions)', () => {
       { delegateType: DelegateTypes.sigAuth, expiresIn: exp }
     )
 
+    let resolved = await resolver.resolve(did)
+    expect(resolved.didDocument).toEqual({
+      '@context': expect.anything(),
+      id: did,
+      verificationMethod: [
+        expect.objectContaining({
+          id: `${did}#controller`,
+          blockchainAccountId: `eip155:1337:${identity}`,
+        }),
+        expect.objectContaining({
+          id: `${did}#delegate-1`,
+          blockchainAccountId: `eip155:1337:${delegate1}`,
+        }),
+      ],
+      authentication: [`${did}#controller`, `${did}#delegate-1`],
+      assertionMethod: [`${did}#controller`, `${did}#delegate-1`],
+    })
+
     // Add second delegate
     const hash2 = await ethrDid.createAddDelegateHash(delegateType, delegate2, exp)
     const signature2 = new SigningKey(currentOwnerPrivateKey).sign(hash2)
@@ -844,6 +864,28 @@ describe('EthrDID (Meta Transactions)', () => {
       },
       { delegateType: DelegateTypes.sigAuth, expiresIn: exp }
     )
+
+    resolved = await resolver.resolve(did)
+    expect(resolved.didDocument).toEqual({
+      '@context': expect.anything(),
+      id: did,
+      verificationMethod: [
+        expect.objectContaining({
+          id: `${did}#controller`,
+          blockchainAccountId: `eip155:1337:${identity}`,
+        }),
+        expect.objectContaining({
+          id: `${did}#delegate-1`,
+          blockchainAccountId: `eip155:1337:${delegate1}`,
+        }),
+        expect.objectContaining({
+          id: `${did}#delegate-2`,
+          blockchainAccountId: `eip155:1337:${delegate2}`,
+        }),
+      ],
+      authentication: [`${did}#controller`, `${did}#delegate-1`, `${did}#delegate-2`],
+      assertionMethod: [`${did}#controller`, `${did}#delegate-1`, `${did}#delegate-2`],
+    })
   })
 
   it('remove delegate1 via meta transaction', async () => {
@@ -855,6 +897,28 @@ describe('EthrDID (Meta Transactions)', () => {
       sigV: signature.v,
       sigR: signature.r,
       sigS: signature.s,
+    })
+
+    // revoking a delegate sets their validity to the block timestamp instead of 0, so we need to wait at least a
+    // second for the resolver to see a difference in the document
+    await sleep(1)
+
+    const resolved = await resolver.resolve(did)
+    expect(resolved.didDocument).toEqual({
+      '@context': expect.anything(),
+      id: did,
+      verificationMethod: [
+        expect.objectContaining({
+          id: `${did}#controller`,
+          blockchainAccountId: `eip155:1337:${identity}`,
+        }),
+        expect.objectContaining({
+          id: `${did}#delegate-2`,
+          blockchainAccountId: `eip155:1337:${delegate2}`,
+        }),
+      ],
+      authentication: [`${did}#controller`, `${did}#delegate-2`],
+      assertionMethod: [`${did}#controller`, `${did}#delegate-2`],
     })
   })
 
@@ -873,6 +937,34 @@ describe('EthrDID (Meta Transactions)', () => {
       sigS: signature1.s,
     })
 
+    let resolved = await resolver.resolve(did)
+    expect(resolved.didDocument).toEqual({
+      '@context': expect.anything(),
+      id: did,
+      verificationMethod: [
+        expect.objectContaining({
+          id: `${did}#controller`,
+          blockchainAccountId: `eip155:1337:${identity}`,
+        }),
+        expect.objectContaining({
+          id: `${did}#delegate-2`,
+          blockchainAccountId: `eip155:1337:${delegate2}`,
+        }),
+      ],
+      authentication: [`${did}#controller`, `${did}#delegate-2`],
+      assertionMethod: [`${did}#controller`, `${did}#delegate-2`],
+      service: [
+        {
+          id: `${did}#service-1`,
+          serviceEndpoint: {
+            transportType: 'http',
+            uri: 'https://didcomm.example.com',
+          },
+          type: 'testService',
+        },
+      ],
+    })
+
     // Add second attribute
     const attributeName2 = 'did/svc/test2Service'
     const hash2 = await ethrDid.createSetAttributeHash(attributeName2, attributeValue, attributeExpiration)
@@ -882,6 +974,42 @@ describe('EthrDID (Meta Transactions)', () => {
       sigV: signature2.v,
       sigR: signature2.r,
       sigS: signature2.s,
+    })
+
+    resolved = await resolver.resolve(did)
+    expect(resolved.didDocument).toEqual({
+      '@context': expect.anything(),
+      id: did,
+      verificationMethod: [
+        expect.objectContaining({
+          id: `${did}#controller`,
+          blockchainAccountId: `eip155:1337:${identity}`,
+        }),
+        expect.objectContaining({
+          id: `${did}#delegate-2`,
+          blockchainAccountId: `eip155:1337:${delegate2}`,
+        }),
+      ],
+      authentication: [`${did}#controller`, `${did}#delegate-2`],
+      assertionMethod: [`${did}#controller`, `${did}#delegate-2`],
+      service: [
+        {
+          id: `${did}#service-1`,
+          serviceEndpoint: {
+            transportType: 'http',
+            uri: 'https://didcomm.example.com',
+          },
+          type: 'testService',
+        },
+        {
+          id: `${did}#service-2`,
+          serviceEndpoint: {
+            transportType: 'http',
+            uri: 'https://didcomm.example.com',
+          },
+          type: 'test2Service',
+        },
+      ],
     })
   })
 
@@ -897,6 +1025,34 @@ describe('EthrDID (Meta Transactions)', () => {
       sigR: signature.r,
       sigS: signature.s,
     })
+
+    const resolved = await resolver.resolve(did)
+    expect(resolved.didDocument).toEqual({
+      '@context': expect.anything(),
+      id: did,
+      verificationMethod: [
+        expect.objectContaining({
+          id: `${did}#controller`,
+          blockchainAccountId: `eip155:1337:${identity}`,
+        }),
+        expect.objectContaining({
+          id: `${did}#delegate-2`,
+          blockchainAccountId: `eip155:1337:${delegate2}`,
+        }),
+      ],
+      authentication: [`${did}#controller`, `${did}#delegate-2`],
+      assertionMethod: [`${did}#controller`, `${did}#delegate-2`],
+      service: [
+        {
+          id: `${did}#service-2`,
+          serviceEndpoint: {
+            transportType: 'http',
+            uri: 'https://didcomm.example.com',
+          },
+          type: 'test2Service',
+        },
+      ],
+    })
   })
 
   it('change owner via meta transaction', async () => {
@@ -910,9 +1066,36 @@ describe('EthrDID (Meta Transactions)', () => {
         sigV: signature.v,
         sigR: signature.r,
         sigS: signature.s,
-      },
-      {}
+      }
     )
+
+    const resolved = await resolver.resolve(did)
+    expect(resolved.didDocument).toEqual({
+      '@context': expect.anything(),
+      id: did,
+      verificationMethod: [
+        expect.objectContaining({
+          id: `${did}#controller`,
+          blockchainAccountId: `eip155:1337:${nextOwner}`,
+        }),
+        expect.objectContaining({
+          id: `${did}#delegate-2`,
+          blockchainAccountId: `eip155:1337:${delegate2}`,
+        }),
+      ],
+      authentication: [`${did}#controller`, `${did}#delegate-2`],
+      assertionMethod: [`${did}#controller`, `${did}#delegate-2`],
+      service: [
+        {
+          id: `${did}#service-2`,
+          serviceEndpoint: {
+            transportType: 'http',
+            uri: 'https://didcomm.example.com',
+          },
+          type: 'test2Service',
+        },
+      ],
+    })
   })
 
   it('resolves document and verify changes', async () => {