+++ src/thirdparty.types.d.ts
@@ -1,2 +1,1 @@
-declare module 'ethr-did-registry'
 declare module 'ganache-cli'